Item Interactions (#5147)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
## ABOUT

This pr is going to replace attackby item interactions with real ones
based on MrMelbert's work
https://github.com/tgstation/tgstation/pull/79968
https://github.com/tgstation/tgstation/pull/83818

Item interactions will take place after tool interactions (which we have
none of) and before regular attacks.

## SURGERY
I changed a bit about surgeries because they were so awful I had to.
(and because their start was moved to an interaction)

Surgeries are started via the same way but now require you to not be in
combat mode and be on weak intent.

Surgeries refactored to use surgeries and updated to latest TG code that
uses limb states, this is practically a better version of what it is
currently, except with a radial instead of a pop-up list.

Intentional surgery failure is done by right click instead of strong
intent to not conflict killing people

Having not enough skill for surgery just simply forces you to fail, so
you can still get experience and be malicious.

Wound cautery changed to a cautery interaction on mobs, the cautery can
still be used to close skin.

Opening organ storage requires weak instead of touch

## INTERACTION CHANGES
You can force feed food to others.

Syringes inject on left click and draw on right click.

Giving items moved to ctrl+shift click or the rebindable keybind "G".
The keybind is reliant on getting a turf from the mouse and doesn't work
when in the process of moving so stand still if you want to use it or
use the static bind.

Slot machine is empty hand right click not attackby

todo for non attack interactions
- [x] attack
- [x] attackby (obvious ones)
- [x] attack_atom (all)
- [x] afterattack (most)
- [x] preattack (most)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Allows atom storage to be better, will cleanup afterattack, preattack,
attackby misuse and attack chain precedence.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Please note that
maintainers freely reserve the right to remove and add tags should they
deem it appropriate. You can attempt to finagle the system all you want,
but it's best to shoot for clear communication right off the bat. -->

<!-- !! Do not add whitespace in-between the entries, do not change the
tags and do not leave the tags without any entries. -->
:cl:
refactor: Refactored attack chain code, report any issues with clicking
on things with other things.
refactor: Refactored surgery code, report any cases of surgery issues.
add: Giving is now done by ctrl+shift click or the hotkey "G"
add: Some interactions have been changed that will not be listed here,
check the PR page before reporting things.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

# Conflicts:
#	code/_onclick/other_mobs.dm
#	code/datums/quirks/peculiarity/_peculiarity.dm
#	code/game/objects/items/books.dm
#	code/game/objects/items/caparison.dm
#	code/game/objects/items/dmusicbox.dm
#	code/game/objects/items/inquisition_relics.dm
#	code/game/objects/items/tools/tongs.dm
#	code/game/objects/structures/fake_machines/ATM.dm
#	code/game/objects/structures/hole.dm
#	code/modules/antagonists/villain/zizo_cult/rituals.dm
#	code/modules/clothing/outfits/patrons.dm
#	code/modules/crafting/alchemy/reagents.dm
#	code/modules/jobs/job_types/_job.dm
#	code/modules/jobs/job_types/adventurer/types/pilgrim/briar.dm
#	code/modules/mob/living/carbon/carbon.dm
#	code/modules/mob/living/carbon/carbon_defense.dm
#	code/modules/projectiles/guns/ballistic/powder/musket.dm
#	code/modules/surgery/surgeries/bestow_lux.dm
#	code/modules/surgery/surgeries/extract_chimeric_node.dm
#	code/modules/surgery/surgeries/extract_lux.dm
#	code/modules/surgery/surgeries/revival.dm
#	html/changelogs/archive/2026-04.yml
#	html/changelogs/archive/2026-06.yml
#	icons/mob/roguehud.dmi
#	icons/obj/structures_spawners.dmi

+++ code/game/objects/items/books.dm
@@ -1120,7 +1120,7 @@
 	fence = TRUE
 
 /obj/item/book/bibble
-	name = "The Motonium"
+	name = "The Book"
 	icon_state = "bibble_0"
 	base_icon_state = "bibble"
 	title = "bible"
@@ -1134,17 +1134,17 @@
 	var/current_verse_range = "Random Verse"
 	var/list/verse_ranges = list(           // Patron divied up by line in txt file
 		"Seperation" = list(1, 4),
-		"Elementalism" = list(5, 9),
-		"Visires" = list(10, 12),
-		"Gani" = list(13, 15),
-		"Valdala" = list(16, 19),
-		"Akan" = list(20, 22),
-		"Goler Kanh" = list(23, 26),
-		"Alaqshir" = list(27, 29),
-		"Pomette" = list(30, 32),
-		"Iliope" = list(33, 36),
-		"Erdl" = list(37, 39),
-		"Mjallidhorn" = list(40, 42),
+		"Heavens" = list(5, 9),
+		"Astrata" = list(10, 12),
+		"Dendor" = list(13, 15),
+		"Necra" = list(16, 19),
+		"Noc" = list(20, 22),
+		"Malum" = list(23, 26),
+		"Ravox" = list(27, 29),
+		"Eora" = list(30, 32),
+		"Xylix" = list(33, 36),
+		"Pestra" = list(37, 39),
+		"Abyssor" = list(40, 42),
 		"Random Verse" = null
 		)
 
@@ -1208,14 +1208,20 @@
 
 	to_chat(user, span_notice("You will read a section from [choice]."))
 
-/obj/item/book/bibble/attack(mob/living/M, mob/user, list/modifiers)
-	if(is_priest_job(user.mind?.assigned_role))
-		if(!user.can_read(src))
-			return
-		M.apply_status_effect(/datum/status_effect/buff/blessed)
-		user.visible_message(span_notice("[user] blesses [M]."))
-		playsound(user, 'sound/magic/bless.ogg', 100, FALSE)
-		return
+/obj/item/book/bibble/interact_with_atom(atom/interacting_with, mob/living/user, list/modifiers)
+	if(!isliving(interacting_with) || !is_priest_job(user.mind?.assigned_role))
+		return NONE
+
+	if(!user.can_read(src))
+		return ITEM_INTERACT_BLOCKING
+
+	var/mob/living/M = interacting_with
+
+	M.apply_status_effect(/datum/status_effect/buff/blessed)
+	user.visible_message(span_notice("[user] blesses [M]."), span_notice("I bless [M]."))
+	playsound(user, 'sound/magic/bless.ogg', 100, FALSE)
+
+	return ITEM_INTERACT_SUCCESS
 
 /datum/status_effect/buff/blessed
 	id = "blessed"
@@ -1254,7 +1260,7 @@
 	base_icon_state = "book5"
 	bookfile = "knowledge.json"
 
-/obj/item/book/secret/iliope
+/obj/item/book/secret/xylix
 	name = "Book of Gold"
 	desc = "{<font color='red'><blink>An ominous book with untold powers.</blink></font>}"
 	icon_state ="xylix_0"
@@ -1263,7 +1269,7 @@
 	base_icon_state = "pellbookmimic"
 	bookfile = "xylix.json"
 
-/obj/item/book/iliope/attack_self(mob/user, list/modifiers)
+/obj/item/book/xylix/attack_self(mob/user, list/modifiers)
 	user.update_inv_hands()
 	to_chat(user, "<span class='notice'>You feel laughter echo in your head.</span>")
 
@@ -1298,7 +1304,7 @@
 
 /obj/item/book/robber
 	name = "Reading for Robbers"
-	desc = "By Flavius of Gani"
+	desc = "By Flavius of Dendor"
 	icon_state ="basic_book_0"
 	base_icon_state = "basic_book"
 	bookfile = "tales4.json"
@@ -1324,23 +1330,23 @@
 	base_icon_state = "book8"
 	bookfile = "tales7.json"
 
-/obj/item/book/mjallidhorn
+/obj/item/book/abyssor
 	name = "A Tale of Those Who Live At Sea"
 	desc = "By Bellum Aegir"
 	icon_state ="book2_0"
 	base_icon_state = "book2"
 	bookfile = "tales8.json"
 
-/obj/item/book/valdala
-	name = "Burial Rites for Valdala"
-	desc = "By Hunlaf, Gravedigger. Revised by Lenore, Priest of Valdala."
+/obj/item/book/necra
+	name = "Burial Rites for Necra"
+	desc = "By Hunlaf, Gravedigger. Revised by Lenore, Priest of Necra."
 	icon_state ="book6_0"
 	base_icon_state = "book6"
 	bookfile = "tales9.json"
 
-/obj/item/book/akan
+/obj/item/book/noc
 	name = "Dreamseeker"
-	desc = "By Hunlaf, Gravedigger. Revised by Lenore, Priest of Valdala."
+	desc = "By Hunlaf, Gravedigger. Revised by Lenore, Priest of Necra."
 	icon_state ="book6_0"
 	base_icon_state = "book6"
 	bookfile = "tales10.json"
@@ -1492,20 +1498,24 @@
 	for(var/obj/item/paper/page as anything in pages)
 		compiled_pages += "<p>[page.info]</p>\n"
 
-/obj/item/manuscript/attackby(obj/item/I, mob/living/user, list/modifiers)
+/obj/item/manuscript/item_interaction(mob/living/user, obj/item/tool, list/modifiers)
+	if(user.cmode)
+		return NONE
+
 	// why is a book crafting kit using the craft system, but crafting a book isn't?
 	// Well, *for some reason*, the crafting system is made in such a way
 	// as to make reworking it to allow you to put reqs vars in the crafted item near *impossible.*
-	if(istype(I, /obj/item/book_crafting_kit))
-		var/obj/item/book/playerbook/PB = new /obj/item/book/playerbook(get_turf(I.loc), TRUE, user, compiled_pages)
-		qdel(I)
+	if(istype(tool, /obj/item/book_crafting_kit))
+		var/obj/item/book/playerbook/PB = new /obj/item/book/playerbook(get_turf(tool), TRUE, user, compiled_pages)
+		qdel(tool)
 		if(user.Adjacent(PB))
 			PB.add_fingerprint(user)
 			user.put_in_hands(PB)
-		return qdel(src)
+		qdel(src)
+		return ITEM_INTERACT_SUCCESS
 
-	if((I.type == /obj/item/paper) || (I.type == /obj/item/paper/scroll))
-		var/obj/item/paper/inserted_paper = I
+	if((tool.type == /obj/item/paper) || (tool.type == /obj/item/paper/scroll))
+		var/obj/item/paper/inserted_paper = tool
 		if(length(pages) == 8)
 			to_chat(user, span_warning("I can not find a place to put [inserted_paper] into [src]..."))
 			return
@@ -1515,8 +1525,41 @@
 		to_chat(user, span_notice("I put [inserted_paper] into [src]."))
 		update_pages()
 		updateUsrDialog()
+		return ITEM_INTERACT_SUCCESS
 
-	return ..()
+/obj/item/manuscript/item_interaction_secondary(mob/living/user, obj/item/tool, list/modifiers)
+	if(user.cmode)
+		return NONE
+
+	if(!istype(tool, /obj/item/natural/feather))
+		return NONE
+
+	if(written)
+		to_chat(user, "<span class='notice'>The manuscript has already been authored and titled.</span>")
+		return ITEM_INTERACT_BLOCKING
+
+	// Prompt user to populate manuscript fields
+	var/newtitle = dd_limittext(SANITIZE_HEAR_MESSAGE(input(user, "Enter the title of the manuscript:") as text|null), MAX_CHARTER_LEN)
+	var/newauthor = dd_limittext(SANITIZE_HEAR_MESSAGE(input(user, "Enter the author's name:") as text|null), MAX_CHARTER_LEN)
+	var/newcategory = input(user, "Select the category of the manuscript:") in list("Apocrypha & Grimoires", "Myths & Tales", "Legends & Accounts", "Thesis", "Eoratica")
+	var/newicon = book_icons[input(user, "Choose a book style", "Book Style") as anything in book_icons]
+
+	if(newtitle && newauthor && newcategory)
+		name = newtitle
+		author = newauthor
+		category = newcategory
+		ckey = user.ckey
+		select_icon = newicon
+		icon_state = "paperwrite"
+		to_chat(user, "<span class='notice'>You have successfully authored and titled the manuscript.</span>")
+		var/complete = tgui_alert(user, "Is the manuscript finished?", "WORDS OF NOC", DEFAULT_INPUT_CHOICES)
+		SEND_SIGNAL(user, COMSIG_BOOK_WRITTEN)
+		if(complete == CHOICE_YES && compiled_pages)
+			written = TRUE
+	else
+		to_chat(user, "<span class='notice'>You must fill out all fields to complete the manuscript.</span>")
+
+	return ITEM_INTERACT_SUCCESS
 
 /obj/item/manuscript/examine(mob/user)
 	. = ..()
@@ -1577,39 +1620,6 @@
 	dat += "<a href='byond://?src=[REF(src)];close=1' style='position:absolute;right:50px'>Close</a>"
 	user << browse(dat, "window=reading;size=1000x700;can_close=1;can_minimize=0;can_maximize=0;can_resize=0;")
 	onclose(user, "reading", src)
-
-
-/obj/item/manuscript/attackby_secondary(obj/item/I, mob/user, list/modifiers)
-	. = ..()
-	if(. == SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN)
-		return
-
-	if(istype(I, /obj/item/natural/feather))
-		if(written)
-			to_chat(user, "<span class='notice'>The manuscript has already been authored and titled.</span>")
-			return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
-		// Prompt user to populate manuscript fields
-		var/newtitle = dd_limittext(SANITIZE_HEAR_MESSAGE(input(user, "Enter the title of the manuscript:") as text|null), MAX_CHARTER_LEN)
-		var/newauthor = dd_limittext(SANITIZE_HEAR_MESSAGE(input(user, "Enter the author's name:") as text|null), MAX_CHARTER_LEN)
-		var/newcategory = input(user, "Select the category of the manuscript:") in list("Apocrypha & Grimoires", "Myths & Tales", "Legends & Accounts", "Thesis", "Heretical")
-		var/newicon = book_icons[input(user, "Choose a book style", "Book Style") as anything in book_icons]
-
-		if(newtitle && newauthor && newcategory)
-			name = newtitle
-			author = newauthor
-			category = newcategory
-			ckey = user.ckey
-			select_icon = newicon
-			icon_state = "paperwrite"
-			to_chat(user, "<span class='notice'>You have successfully authored and titled the manuscript.</span>")
-			var/complete = tgui_alert(user, "Is the manuscript finished?", "WORDS OF AKAN", DEFAULT_INPUT_CHOICES)
-			SEND_SIGNAL(user, COMSIG_BOOK_WRITTEN)
-			if(complete == CHOICE_YES && compiled_pages)
-				written = TRUE
-		else
-			to_chat(user, "<span class='notice'>You must fill out all fields to complete the manuscript.</span>")
-
-		return SECONDARY_ATTACK_CANCEL_ATTACK_CHAIN
 
 /obj/item/manuscript/update_icon_state()
 	. = ..()
@@ -1683,8 +1693,8 @@ ____________End of Example*/
 	base_icon_state = "knowledge"
 	bookfile = "MagicalTheory.json"
 
-/obj/item/book/vowvaldalapage
-	name = "Valdala's Vow of Silence"
+/obj/item/book/vownecrapage
+	name = "Necra's Vow of Silence"
 	desc = "A faded page, with seemingly no author."
 	icon_state = "book8_0"
 	base_icon_state = "book8"
@@ -1736,22 +1746,28 @@ ____________End of Example*/
 	bookfile = "Neu_cooking.json"
 
 /obj/item/book/bibble/psy
-	name = "Angrosian Motonium"
+	name = "The PSY Book"
 	icon_state = "bibble_0" // change when sprites avaliable
 	base_icon_state = "bibble"
-	title = "angrosian bible"
+	title = "psydon bible"
 	dat = "gott.json"
 	verses_file = "strings/psybibble.txt"
 	can_select_verse_ranges = FALSE
 
-/obj/item/book/bibble/psy/attack(mob/living/M, mob/living/user, list/modifiers)
-	if(istype(user) && istype(user.patron, /datum/patron/angros))
-		if(!user.can_read(src))
-			return
-		M.apply_status_effect(/datum/status_effect/buff/blessed)
-		user.visible_message(span_notice("[user] blesses [M]."))
-		playsound(user, 'sound/magic/bless.ogg', 100, FALSE)
-		return
+/obj/item/book/bibble/psy/interact_with_atom(atom/interacting_with, mob/living/user, list/modifiers)
+	if(!isliving(interacting_with) || !istype(user.patron, /datum/patron/psydon))
+		return NONE
+
+	if(!user.can_read(src))
+		return ITEM_INTERACT_BLOCKING
+
+	var/mob/living/M = interacting_with
+
+	M.apply_status_effect(/datum/status_effect/buff/blessed)
+	user.visible_message(span_notice("[user] blesses [M]."), span_notice("I bless [M]."))
+	playsound(user, 'sound/magic/bless.ogg', 100, FALSE)
+
+	return ITEM_INTERACT_SUCCESS
 
 /datum/status_effect/buff/blessed
 	id = "blessed"
